+++ mpu/cli.py
@@ -20,17 +20,17 @@ __version__ = "0.4.2"
 def getstock(
         current_price_strategy: CurrentPriceStrat,
         price_update_strategy: PriceUpdaterStrat,
-        strategies_options_path: Path = typer.Option(
-            None,
-            "--strategies-options-path",
-            "-sop",
+        config_path: Path = typer.Option(
+            ...,
+            "--config-path",
+            "-cp",
             exists=True,
             file_okay=True,
             dir_okay=True,
             writable=True,
             readable=True,
             resolve_path=True,
-            help="Path of the file to configure the strategies, if not provided not parameters are used",
+            help="Path of the file to configure mpu",
         ),
         market_extract_path: Path = typer.Option(
             lambda: Path.cwd(),
@@ -69,7 +69,7 @@ def getstock(
     main_getstock(
         current_price_strategy=current_price_strategy.value,
         price_update_strategy=price_update_strategy.value,
-        strategies_options_path=strategies_options_path,
+        config_path=config_path,
         market_extract_path=market_extract_path,
         output_path=output_path,
         force_update=force_download,